+++ CHANGELOG.rst
@@ -79,3 +79,4 @@ latest
 * Clarify behaviour of get_import_details.
 * Add module_is_squashed method.
 * Add squash_module method.
+* Add find_all_simple_chains method.